feat(lib): add possibility to blacklist package installation
This is done by overriding the installed package to a dummy package.
This commit additionally introduces this change to the 'zsh' module.

Note that this is not possible for every package, but rather for ones
that use simple installation processes or ones that do not generate
nix-store dependent file paths.